+++ app/src/main/java/com/fortitude/shamsulkarim/ieltsfordory/NewWordFragment.java
@@ -278,47 +278,47 @@ public class NewWordFragment extends Fragment {
 
         words.clear();
 
-        double IELTSwordSize = getResources().getStringArray(R.array.IELTS_words).length;
-        double TOEFLwordSize = getResources().getStringArray(R.array.TOEFL_words).length;
-        double SATwordSize = getResources().getStringArray(R.array.SAT_words).length;
-        double GREwordSize = getResources().getStringArray(R.array.GRE_words).length;
+        int IELTSwordSize = getResources().getStringArray(R.array.IELTS_words).length;
+        int TOEFLwordSize = getResources().getStringArray(R.array.TOEFL_words).length;
+        int SATwordSize = getResources().getStringArray(R.array.SAT_words).length;
+        int GREwordSize = getResources().getStringArray(R.array.GRE_words).length;
 
-        double IELTSbeginnerNumber = 0;
-        double TOEFLbeginnerNumber = 0;
-        double SATbeginnerNumber = 0;
-        double GREbeginnerNumber = 0;
+        int IELTSbeginnerNumber = 0;
+        int TOEFLbeginnerNumber = 0;
+        int SATbeginnerNumber = 0;
+        int GREbeginnerNumber = 0;
 
 
 
 
         if(isIeltsChecked){
-            IELTSbeginnerNumber = getPercentageNumber(30d, IELTSwordSize);
+            IELTSbeginnerNumber = (int)getPercentageNumber(30, IELTSwordSize);
 
         }
 
         if(isToeflChecked){
 
-            TOEFLbeginnerNumber = getPercentageNumber(30d, TOEFLwordSize);
+            TOEFLbeginnerNumber = (int) getPercentageNumber(30, TOEFLwordSize);
 
         }
 
         if(isSatChecked){
 
-            SATbeginnerNumber = getPercentageNumber(30d, SATwordSize);
+            SATbeginnerNumber = (int) getPercentageNumber(30, SATwordSize);
 
         }
 
         if(isGreChecked){
 
 
-            GREbeginnerNumber = getPercentageNumber(30d, GREwordSize);
+            GREbeginnerNumber = (int) getPercentageNumber(30, GREwordSize);
 
         }
 
-        addIELTSwords(0d,IELTSbeginnerNumber);
-        addTOEFLwords(0d,TOEFLbeginnerNumber);
-        addSATwords(0d,SATbeginnerNumber);
-        addGREwords(0d,GREbeginnerNumber);
+        addIELTSwords(0,IELTSbeginnerNumber);
+        addTOEFLwords(0,TOEFLbeginnerNumber);
+        addSATwords(0,SATbeginnerNumber);
+        addGREwords(0,GREbeginnerNumber);
 
 
 
@@ -331,49 +331,49 @@ public class NewWordFragment extends Fragment {
 
         words.clear();
 
-        double IELTSwordSize = getResources().getStringArray(R.array.IELTS_words).length;
-        double TOEFLwordSize = getResources().getStringArray(R.array.TOEFL_words).length;
-        double SATwordSize = getResources().getStringArray(R.array.SAT_words).length;
-        double GREwordSize = getResources().getStringArray(R.array.GRE_words).length;
+        int IELTSwordSize = getResources().getStringArray(R.array.IELTS_words).length;
+        int TOEFLwordSize = getResources().getStringArray(R.array.TOEFL_words).length;
+        int SATwordSize = getResources().getStringArray(R.array.SAT_words).length;
+        int GREwordSize = getResources().getStringArray(R.array.GRE_words).length;
 
-        double IELTSintermediateNumber = 0;
-        double TOEFLintermediateNumber = 0;
-        double SATintermediateNumber = 0;
-        double GREintermediateNumber = 0;
+        int IELTSintermediateNumber = 0;
+        int TOEFLintermediateNumber = 0;
+        int SATintermediateNumber = 0;
+        int GREintermediateNumber = 0;
 
-        double IELTSbeginnerNumber = 0;
-        double TOEFLbeginnerNumber = 0;
-        double SATbeginnerNumber = 0;
-        double GREbeginnerNumber = 0;
+        int IELTSbeginnerNumber = 0;
+        int TOEFLbeginnerNumber = 0;
+        int SATbeginnerNumber = 0;
+        int GREbeginnerNumber = 0;
 
 
 
 
         if(isIeltsChecked){
-            IELTSintermediateNumber = getPercentageNumber(40d, IELTSwordSize);
-            IELTSbeginnerNumber = getPercentageNumber(30d, IELTSwordSize);
+            IELTSintermediateNumber = getPercentageNumber(40, IELTSwordSize);
+            IELTSbeginnerNumber = getPercentageNumber(30, IELTSwordSize);
 
         }
 
         if(isToeflChecked){
 
-            TOEFLintermediateNumber = getPercentageNumber(40d, TOEFLwordSize);
-            TOEFLbeginnerNumber = getPercentageNumber(30d, TOEFLwordSize);
+            TOEFLintermediateNumber = getPercentageNumber(40, TOEFLwordSize);
+            TOEFLbeginnerNumber = getPercentageNumber(30, TOEFLwordSize);
 
         }
 
         if(isSatChecked){
 
-            SATintermediateNumber = getPercentageNumber(40d, SATwordSize);
-            SATbeginnerNumber = getPercentageNumber(30d, SATwordSize);
+            SATintermediateNumber = getPercentageNumber(40, SATwordSize);
+            SATbeginnerNumber = getPercentageNumber(30, SATwordSize);
 
         }
 
         if(isGreChecked){
 
 
-            GREintermediateNumber = getPercentageNumber(40d, GREwordSize);
-            GREbeginnerNumber = getPercentageNumber(30d, GREwordSize);
+            GREintermediateNumber = getPercentageNumber(40, GREwordSize);
+            GREbeginnerNumber = getPercentageNumber(30, GREwordSize);
 
         }
 
@@ -390,49 +390,49 @@ public class NewWordFragment extends Fragment {
 
         words.clear();
 
-        double IELTSwordSize = getResources().getStringArray(R.array.IELTS_words).length;
-        double TOEFLwordSize = getResources().getStringArray(R.array.TOEFL_words).length;
-        double SATwordSize = getResources().getStringArray(R.array.SAT_words).length;
-        double GREwordSize = getResources().getStringArray(R.array.GRE_words).length;
+        int IELTSwordSize = getResources().getStringArray(R.array.IELTS_words).length;
+        int TOEFLwordSize = getResources().getStringArray(R.array.TOEFL_words).length;
+        int SATwordSize = getResources().getStringArray(R.array.SAT_words).length;
+        int GREwordSize = getResources().getStringArray(R.array.GRE_words).length;
 
-        double IELTSintermediateNumber = 0;
-        double TOEFLintermediateNumber = 0;
-        double SATintermediateNumber = 0;
-        double GREintermediateNumber = 0;
+        int IELTSintermediateNumber = 0;
+        int TOEFLintermediateNumber = 0;
+        int SATintermediateNumber = 0;
+        int GREintermediateNumber = 0;
 
-        double IELTSbeginnerNumber = 0;
-        double TOEFLbeginnerNumber = 0;
-        double SATbeginnerNumber = 0;
-        double GREbeginnerNumber = 0;
+        int IELTSbeginnerNumber = 0;
+        int TOEFLbeginnerNumber = 0;
+        int SATbeginnerNumber = 0;
+        int GREbeginnerNumber = 0;
 
 
 
 
         if(isIeltsChecked){
-            IELTSintermediateNumber = getPercentageNumber(40d, IELTSwordSize);
-            IELTSbeginnerNumber = getPercentageNumber(30d, IELTSwordSize);
+            IELTSintermediateNumber = getPercentageNumber(40, IELTSwordSize);
+            IELTSbeginnerNumber = getPercentageNumber(30, IELTSwordSize);
 
         }
 
         if(isToeflChecked){
 
-            TOEFLintermediateNumber = getPercentageNumber(40d, TOEFLwordSize);
-            TOEFLbeginnerNumber = getPercentageNumber(30d, TOEFLwordSize);
+            TOEFLintermediateNumber = getPercentageNumber(40, TOEFLwordSize);
+            TOEFLbeginnerNumber = getPercentageNumber(30, TOEFLwordSize);
 
         }
 
         if(isSatChecked){
 
-            SATintermediateNumber = getPercentageNumber(40d, SATwordSize);
-            SATbeginnerNumber = getPercentageNumber(30d, SATwordSize);
+            SATintermediateNumber = getPercentageNumber(40, SATwordSize);
+            SATbeginnerNumber = getPercentageNumber(30, SATwordSize);
 
         }
 
         if(isGreChecked){
 
 
-            GREintermediateNumber = getPercentageNumber(40d, GREwordSize);
-            GREbeginnerNumber = getPercentageNumber(30d, GREwordSize);
+            GREintermediateNumber = getPercentageNumber(40, GREwordSize);
+            GREbeginnerNumber = getPercentageNumber(30, GREwordSize);
 
         }
 
@@ -447,17 +447,17 @@ public class NewWordFragment extends Fragment {
     }
 
 
-    private double getPercentageNumber(double percentage, double number) {
+    private int getPercentageNumber(int percentage, int number) {
 
 
         double p = percentage / 100d;
         double beginnerNum = p * number;
 
-        return beginnerNum;
+        return (int)beginnerNum;
 
     }
 
-    private  void addIELTSwords(double startPoint,double IELTSbeginnerNumber){
+    private  void addIELTSwords(int startPoint,int IELTSbeginnerNumber){
 
 
 
@@ -476,7 +476,7 @@ public class NewWordFragment extends Fragment {
 
     }
 
-    private  void addTOEFLwords(double startPoint, double TOEFLbeginnerNumber){
+    private  void addTOEFLwords(int startPoint, int TOEFLbeginnerNumber){
 
 
         if(isToeflChecked){
@@ -495,7 +495,7 @@ public class NewWordFragment extends Fragment {
 
     }
 
-    private void addSATwords (double startPoint ,double SATbeginnerNumber){
+    private void addSATwords (int startPoint ,int SATbeginnerNumber){
 
 
         if(isSatChecked){
@@ -516,7 +516,7 @@ public class NewWordFragment extends Fragment {
 
     }
 
-    private void addGREwords (double startPoint ,double SATbeginnerNumber){
+    private void addGREwords (int startPoint ,int SATbeginnerNumber){
 
 
         if(isGreChecked){
